Adjusted modifications table migration for deployment

+++ database/migrations/2024_05_14_014152_create_modifications_table.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
@@ -6,25 +6,27 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-     /**
+    /**
      * Run the migrations.
      */
     public function up()
     {
-        Schema::create('modifications', function (Blueprint $table) {
-            $table->id();
-            $table->foreignIdFor(\App\Models\Build::class)->constrained()->onDelete('cascade');
-            $table->string('category');
-            $table->string('name');
-            $table->string('brand')->nullable();
-            $table->decimal('price', 8, 2)->nullable();
-            $table->string('part')->nullable();
-            $table->text('notes')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('modifications')) {
+            Schema::create('modifications', function (Blueprint $table) {
+                $table->id();
+                $table->foreignIdFor(\App\Models\Build::class)->constrained()->onDelete('cascade');
+                $table->string('category');
+                $table->string('name');
+                $table->string('brand')->nullable();
+                $table->decimal('price', 8, 2)->nullable();
+                $table->string('part')->nullable();
+                $table->text('notes')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
-     /**
+    /**
      * Reverse the migrations.
      */
     public function down()
